Upgrade

+++ Contest/TheMillionthFibonacciKata/Tyndraxis/Fibonacci.cs
@@ -8,13 +8,13 @@ public class Fibonacci
     private static readonly double secondFi = (1 - Math.Sqrt(5)) / 2;
 
     public static BigInteger Fib(int n)
-    {
+    { 
         var modifier = n >= 0
             ? 1L
             : n % 2 > 0
                 ? -1L
                 : 1;
-        return new BigInteger(modifier * (long) Math.Round(SolveFibonacci(n)));
+        return new BigInteger(modifier * Math.Round(SolveFibonacci(n)));
     }
 
     private static double SolveFibonacci(int n)
